+++ demos/line/ordinaryPolyline.jsx
@@ -1,45 +1,17 @@
 import React from 'react';
 import { Chart, Line, Point } from '../../src';
 
+let year = 1990;
+
+
 // 数据源
-const data = [
-      {
-        year: "1991",
-        value: 3
-      },
-      {
-        year: "1992",
-        value: 4
-      },
-      {
-        year: "1993",
-        value: 3.5
-      },
-      {
-        year: "1994",
-        value: 5
-      },
-      {
-        year: "1995",
-        value: 4.9
-      },
-      {
-        year: "1996",
-        value: 6
-      },
-      {
-        year: "1997",
-        value: 7
-      },
-      {
-        year: "1998",
-        value: 9
-      },
-      {
-        year: "1999",
-        value: 13
-      }
-    ];
+const data = new Array(1200).fill('').map(() => {
+  return {
+    year: year++,
+    value: Math.round(Math.random())
+  }
+});
+
 
 function Demo() {
   return <Chart
@@ -47,6 +19,7 @@ function Demo() {
     autoFit
     height={500}
     data={data}
+    notCompareData
     scale={{ value: { min: 0 } }}
   >
     <Line position="year*value"  />