+++ unittest/exports/theme-spec.tsx
@@ -1,5 +1,5 @@
 import { registerShape, registerTheme, getTheme } from '../../src/core';
-import { registerPlotTheme } from '../../src';
+// import { registerPlotTheme } from '../../src';
 
 
 describe('export function', () => {
@@ -12,8 +12,8 @@ describe('export function', () => {
   it('should export getTheme', () => {
     expect(getTheme).toBeDefined();
   });
-  it('should export registerPlotTheme', () => {
-    expect(registerPlotTheme).toBeDefined();
-  });
+  // it('should export registerPlotTheme', () => {
+  //   expect(registerPlotTheme).toBeDefined();
+  // });
 })